+++ src/modules/sbstudio/plugin/operators/__init__.py
@@ -16,6 +16,7 @@ from .duplicate_light_effect import DuplicateLightEffectOperator
 from .export_to_csv import SkybrushCSVExportOperator
 from .export_to_dac import DACExportOperator
 from .export_to_dss import DSSPathExportOperator, DSSPath3ExportOperator
+from .export_to_evsky import EVSKYExportOperator
 from .export_to_drotek import DrotekExportOperator
 from .export_to_litebee import LitebeeExportOperator
 from .export_to_skyc import SkybrushExportOperator
@@ -79,6 +80,7 @@ from .remove_schedule_override_entry import RemoveScheduleOverrideEntryOperator
 from .remove_storyboard_entry import RemoveStoryboardEntryOperator
 from .reorder_formation_markers import ReorderFormationMarkersOperator
 from .return_to_home import ReturnToHomeOperator
+from .run_full_proximity_check import RunFullProximityCheckOperator
 from .select_formation import SelectFormationOperator, DeselectFormationOperator
 from .select_storyboard_entry import SelectStoryboardEntryForCurrentFrameOperator
 from .set_server_url import SetServerURLOperator
@@ -105,6 +107,7 @@ __all__ = (
     "DSSPathExportOperator",
     "DSSPath3ExportOperator",
     "DuplicateLightEffectOperator",
+    "EVSKYExportOperator",
     "FixConstraintOrderingOperator",
     "AddMarkersFromQRCodeOperator",
     "GetFormationStatisticsOperator",
@@ -123,6 +126,7 @@ __all__ = (
     "RemoveStoryboardEntryOperator",
     "ReorderFormationMarkersOperator",
     "ReturnToHomeOperator",
+    "RunFullProximityCheckOperator",
     "SelectFormationOperator",
     "SelectStoryboardEntryForCurrentFrameOperator",
     "SetServerURLOperator",